platform: allow env-overrides in base64
Currently we parse env-overrides request headers with `,` to separate
between different pair of variable's name and value. This approach works
for simple value string but when folks want to pass a JSON containing
commas characters within, the parse would fail.

Provide `env-overrides-base64` as a work around for more complex values.
User could encode the `k=v` pair in base64 and pass it to the new
header. Our server will attempt to decode the pair and append it to
existing overrides.