Use GetUser to communicate if user can't access selected group.
Use it to communicate if the user has selected a group that they have
access to but cannot access due to IP rules. Also use the same mechanism
for communicating subdomain access issues.

Update UI to show a minimal sidebar if user doesn't have access to the
selceted group so that the user still has access to the group switcher.