bazelrc: remote-target-linux use platform in output dir, stop repo RE
This flag changes the name of output directories under `bazel-out/*` to
contain the platform information. This means that if a user were to
switch between a local build and a remote build
using `--config=remote-target-linux`, they would be able to retain the
output from both configs.

Also disabling `experimental_repo_remote_exec`. This flag is meant to
allow repository rules, when marked with `remotable = True` to be
executed remotely with RBE. The main purpose of this is to allow
toolchains to probe remote execution platforms information to setup
appropriate toolchains and paths. We do not use this effect in our
repository anywhere, however it will make us re-run all repository rules
when toggle between `--config=remote-target-linux` build and local build.