distributed: fix potential panic
When replication factor is higher than the actual peer count, such as
running 1 pod with replication factor set to 3, we would run into:

```
panic: runtime error: slice bounds out of range [:3] with capacity 0
```

To avoid this, let's make sure that we only sub-slicing when the
replication factor was set correctly.

Also adjust the default replication factor to be inline with our
documentation "Must be >=1".