Make impersonation play nice with subdomains.
If the user is impersonating group X, but tries to access subdomain for group Y then don't enable impersonation on requests for subdomain Y so that subdomain Y can be accessed normally.

This PR slightly changes the semantics of how the frontend communicates impersonation information to the backend. Prior to this change, the frontend would set an impersonation target and would assume that impersonation was in effect.
Now, the frontend still sets the impersonation target but now the backend informs the frontend whether impersonation is actually in effect via the GetUser response.

Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/2779