gossip: fix flaky and race test
It seems like the timeout on these tests could be a bit low when they
are scheduled on resource limitted Executors.

Upon crash, the test will also trigger race detection of accessing the
variable during `t.Fatalf()` while writing to it in an unfinished
goroutine. Avoid the race detection by acquiring the mutex prior to
`t.Fatalf()`.