platforms: initial platformization
Convert usages of `exec_properties` in `//server` into concrete
platforms.

We achieve that by introducing `//toolchains` and `//platforms`.

- `//toolchains` contains the toolchain that we will be using for
  execution.

- `//platforms` contains all the target and exec platforms definitions
  we use.

Additionally, introduced `//platforms/config_settings` to hold all the
config_setting and related config_values.

With this, we only need to declare build/test targets with either
`exec_compatible_with` or `target_compatible_with` using the
config_values knob that we are introducing.

The current `config_{setting,value}` are not finalized. In the long
term, we may want to separate these platforms/config/toolchain
definition to a separate repo for folks to consume.