rules_go: fix GOPROXY unset
In Go 1.21, Go SDK starts to include `go.env` file at the root of the
toolchain with GOPROXY set to default values. With this changes, Go
toolings start to fail when the `go.env` file is not found and the proxy
is unset.

We actually don't get this problem when using Linux RBE because our
action execution image has Go pre-installed and there seems to be some
magic fall back somewhere. However when attempting to build using MacOS
RBE, the error occurred on GoToolchainBinaryBuild actions.

Luckily, this was reported in Bazel slack and fixed in (1).
This PR upgrade our rules_go to that exact fixed commit.

(1): https://github.com/bazelbuild/rules_go/pull/3722