DB API refresh, part 1
This PR defines a slightly more opinionated API for interacting with the database and requires names for all queries.

There's an escape hatch available to fall back to using the raw GORM API.

Part 1 updates all the transaction usages and a few non-transaction usages.

Part 2 will update the remaining queries.