Fix flakiness in TestKeyReencryption.
There was a race condition between the key refresh goroutine and the test advancing the timer.
Prior to this change, there was no guarantee that the refresh goroutine was waiting on the timer at the time the test advanced the time which would lead to the test waiting indefinitely.

Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/2967