macos: make --config=remote work again
In a previous change, we introduced --linkopt unconditionally on MacOS.
This cause remote builds with `--config=remote` to include those linker
flags in our Linux actions. And `ld.gold` did not like these Xcode 15
only flags.

Fix this by making sure that the linker flags are only applied when this
combination happen:
- We are building on MacOS. This is done by using the `common:macos`
  profile that is automatically included by Bazel's
  `--enable_platform_specific_config` flag.

- We are targeting the local MacOS platform. The local platform is
  `@local_config_platform//:host`. We leverage Bazel's platform_mapping
  feature to achieve this.