buildpatches: replace module references with alias
Because of the visibility issue in bzlmod (1), repos that are
generated from a module extension cannot depend on build targets in
first-level modules directly.

The workaround is to create an alias of these first-level module's
targets in the main repository, then use the special alias `@@` to refer
to the main repository.

We expect this issue to be fixed in Bazel 7.2.

(1):