nogo: enable SA4011
```
ineffective break statement. Did you mean to break out of the outer loop?
```

Most of the cases, `break` statement is not needed as it simply breakout
of the `select` clause.

However, there are some cases where `break` was misinterpret as breaking
out of the outer loop, or exit the function. Fixed those cases to use
either `break {label}` or `return` instead.