nogo: enable SA4009
In filepath.WalkDir, the WalkFunc has an err argument that we forgot to
handle.

From https://pkg.go.dev/path/filepath#WalkFunc

```
The err argument reports an error related to path, signaling that Walk
will not walk into that directory. The function can decide how to handle
that error; as described earlier, returning the error will cause Walk to
stop walking the entire tree.

Walk calls the function with a non-nil err argument in two cases.

First, if an os.Lstat on the root directory or any directory or file in
the tree fails, Walk calls the function with path set to that directory
or file's path, info set to nil, and err set to the error from os.Lstat.

Second, if a directory's Readdirnames method fails, Walk calls the
function with path set to the directory's path, info, set to an
fs.FileInfo describing the directory, and err set to the error from
Readdirnames.
```

Since we are not doing any partial walk to make use of filepath.SkipDir
or filepath.SkipAll, returning the error directly to stop the entire
walk.