remotebazel: replacing go-git with git shell
When calling to the remote repository that was setup with ssh protocol,
go-git would require additional auth configuration which might be harder
to obtain.

Replace go-git with plain "git" solves the problem as user should
already have authentication configured for git setup locally.

This also enables the usage of `ls-remote --symref`, which resolves the
symbolic refs on the remote origin. Since the only symbolic ref exists on
the remote side is HEAD, we can assume that this flag provide us the
correct default branch of the repo on the remote side.

Replaces #6362