timing_card: get profile from structured flag
Currently we are assuming that the BuildToolLogs event will have
relatively small amount of files that are uploaded via bytestream API.
Using that assumption, we guessed that the build timing profile must be
one of the 2 files referenced in this event, and differentiate them
based on their file name.

Detect the actual timing profile name by searching through canonical
command line for a `--profile=` flag. If the user does not specify a
custom profile name, assume the default `command.profile.gz` name.