Add target and mnemonic to Execution
Each Execution is often corresponding to a Bazel's Action, but end users
don't often think of Execution or Action because Bazel's build rules
system often abstract that away from them.

Instead, folks often think of Bazel build targets.
Each targets uses a build rule, in combination with a build
configuration, to create multiple actions. Usually each of these actions
are created with a unique mnemonic to help identify them a bit easier
(note that this is not always the case).

When create the Execution, also try to extract the combination of
`TargetID + Mnemonic + ConfigurationID` to help identify the action.
These information should help end users search the relevant actions a
bit easier on our WebUI.