action card timing: ensure the min css value
In #6576, we made sure that all duration in the Timeline phases are
non-negative when time drift happens.

However, because the total duration is still being calculated by taking
the difference between the first event and last event, which is
inaccurate because the first event timestamp could be after the second
event timestamp due to drift. As the total duration is shorter than it
should be, it's possible for the `offset% + event%` value to be
higher than the actual total duration (100%). This causes the last bar of
the timeline, which is calculated with `100% - (offset% + event%)`, to
have a negative value. This prevents `flex` from not being rendered in
the div's style due to invalid initial value, causing strange offset.

Apply several fixes to mitigate the issue:

- Fix total duration calculation by taking the difference between
  the smallest timestamp and the longest timestamp.

- Refactor the code to only add to offset after the flex initial values
  are calculated.

- Ensure that flex initial values cannot goes below zero.

- Fix "Timeline" still being rendered when no event is available.

Closes https://github.com/buildbuddy-io/buildbuddy-internal/issues/3377