dirtools: refactor UploadTree to us fs.WalkDir
Replace the complicated recursion setup with a more straightforward
fs.WalkDir implementation. Files are now eagerly queued for upload as
the traversal happen, while directories and trees are queued after the
tree traversal to ensure the merkle tree computed correctly.

This comes with a side effect: the ordering of "repeated" fields in
remote execution protos, such as Tree.Children or Directory.Files, could
change. The spec recommends that we should sort these fields using
lexical ordering wherever possible, or a stable ordering in other cases.
Effectively, we see that the Tree object in TestUploadTree/LotsOfNesting
needed to be updated because the ordering in Tree.Root.Directory and
Tree.Children was updated. For this reason, added stable sorting on the
relevant slices to make sure that future updates are backward compatible.