event_index: drop Fetch event
Folks who are interested in Fetch information don't simply want a binary
success/fail status, but they also want to check the execution of the
related Bazel repository rules as a whole.

Since that information is often available via the Timing profile with
great resolution, displaying this event in RAW tab is a waste of space.