cli/printlog: provide a --raw flag to print compact exec log
Bazel team has commented that the compact log file format could be
unoptimized for `bazel test //...` where test targets could introduce a
lot of files which cannot be deduplicated.

Provide a flag to inspect the raw content of the compact exec log
to help aid development of this log format.