.bazelrc: write exec log on Windows
This was previously disabled on Windows because we were writing it to
`/tmp` and on Windows, there is not a fixed tmp path.

We have since switched the log path to our workspace and thus,
compatible with Windows.

Also changed the log to use the correct ZSTD extension, which is `.zst`.