remote_asset_test: add test to validate http proxy is used if set
As we start to set

```golang
&http.Transport{
    ...
    Proxy: http.ProxyFromEnvironment,
}
```

it's worth to note that http.ProxyFromEnvironment is initialized only
once and store the proxy result in the package level without a way to
reset it.

So testing this in fetch_server_test.go was not possible, unless we
write our test func at the very top of the file, or shard the test
enough so that each test func gets their own shard.

Instead, implement the test in our integration test setup.