fetch_server: support storing blob using checksum digest func
Originally suggested by @hunshcn in #6911.

In each FetchBlobRequest, there are 2 digest functions:

a. One digest function to verify the content of the downloaded blob.
This is usually set by the Bazel repository rules calling
`repository_ctx.download` and `repository_ctx.download_and_extract` API.
Usually, this is set by http_archive rules to SHA256, but in recent
community rules, such as rules_js and rules_proto, SHA512 and SHA384 are
also used.

b. One digest function to store the blob in cache for the client to
subsequently downloaded. This is usually set to either SHA1, SHA256, or
Blake3, which are the digest functions Bazel support in `startup
--digest_function`. The default is SHA256.

If the Bazel client is also configured with
`--experimental_remote_downloader` pointing to BuildBuddy, client would
get no cache hits from BuildBuddy side if (a) and (b) differs. This is
a result of FetchServer storing the downloaded blob using (b) while
looking up the cache using (a) before downloading.

In #6924, we added support for using these new digest functions in our
cache storage layer.

In #6911, we added support for verifying downloaded blobs using these
new digest functions to Fetch Server.

This PR added the ability to store downloaded blob in our cache using
(a), in additional to existing entry stored with (b). It allows
subsequent Fetch Requests to get cache hits from BuildBuddy at the cost
of storing the blob twice.