invocation_exec_log_card: fix invalid href local runner entries
While working on #7166, I discovered that clicking on `local` actions in
the execution log card would crash our UI.

This is because local spawn entry does not have `digest` and thus,
render the entries in the card with `href=/invocation/<uuid>$#action`,
and clicking on it would cause our browser to crash.

Don't render the invalid URL here.