invocation_exec_log_card: fix sort direction radio group
This was using the same name as the radio buttons for the remote
executions card.
Changing sort direction when both 4 radio buttons are available cause
the checked status to not render correctly.

Rename the radio inputs of exec log.
Also tighten the onChange handler to not rely on the input name to
determine the state.

Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/3587