Reapply "website: add support to run from within the directory (#7399)" (#7421)
Don't add the workspace directory in DEFAULT_CMD_TPL as it's executed
inside a build action and won't have BUILD_WORKSPACE_DIRECTORY
available.

Instead of using the variable directly, use a proxy `WS_DIR` variable so
that if BUILD_WORKSPACE_DIRECTORY is unset, we just gona pass and empty
string to `cd`.

Also replaced all the `&&` with `set -e` to simplify the script.