Add execution ID to ctx sooner (#7533)
We have a few places where Action.command_digest is apparently nil. Add
the execution ID to ctx sooner in one of these cases to help debug
further (the execution ID has the action_digest, so we'll be able to see
whether the command_digest was actually nil, or if we have some bug).