reapi: favor platform set in Action over Command
In REAPI v2.2, the usage of platform in Command was deprecated over in
Action. Although most of the older REAPI clients, such as Bazel, are
setting it in both, newer clients such as Buck2 would only set it in
Action message.

This change replace most of existing `command.GetPlatform()` calls with
`action.GetPlatform()` while using the command value as fallback.

Closes https://github.com/buildbuddy-io/buildbuddy-internal/issues/3863