Proxy Kythe apis to codesearch server (#7546)
1) Build ts protos of kythe stuff we're going to use in the UI.
2) Add a util/grpc_forward library called from grpc_server.go that
forwards traffic, only if forwarding is configured, to some backend
server.

I considered just manually implementing these methods in
buildbuddy_service and calling the kythe backend service (writing out
the proxy basically), but there are between 10 and 24 methods we'd want
to proxy to the kythe backend, so this seemed cleaner and simpler. It
also allows us to proxy other stuff in the future (like the codesearch
APIs, for example).

LMK what you think, I'm open.