Fix --remote_header propagation in CLI local cache proxy (#7571)
Fixes a small regression from
https://github.com/buildbuddy-io/buildbuddy/pull/7563

The PR restructured some things so that we can use `log.Ctx*` to get
more helpful logs. But it introduced a bug - the previous code was
correctly propagating _all_ gRPC metadata to the remote cache (i.e. all
`--remote_header` values) but that PR made it so that we only propagate
the metadata allowed by the RPC interceptors (API key, bazel
requestmetadata, etc.)