Move workflow prober to the probers org (#7619)
Workflows require you to link a GitHub repo. A GitHub org can only be
linked to a single BB org, so we had to link the workflow prober to the
BB org which unfortunately clogged up the build history.

Switch the prober to use remote bazel, which doesn't require a linked
GitHub repo, so we can move the prober to the prober org.


<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**:
https://github.com/buildbuddy-io/buildbuddy-internal/pull/3915