Ensure that worker_completed_timestamp is set (#7658)
As of github.com/buildbuddy-io/buildbuddy/pull/7596, when executions
fail due to an error (and not the action failing),
worker_start_timestamp is set, but worker_finish_timestamp is 0, which
results in large negative durations.

Always set the completed time to at least the start time, and possibly
some larger value. This will start counting usage for errored
executions.