Remove flakiness in remote_execution_test.TestActionMerging_Hedging
In the previous version, it was possible for the 2 slow executions to be accepted and in the executing state, but not actually check for the existence of fname before the unlock command ran.

Now, there is synchronization between the scripts, so unlock will only run after the first 2 executions checked for the existence of fname.