reapi: favor platform set in Action over Command
In REAPI v2.2, the usage of platform in Command was deprecated over in
Action. Although most of the older REAPI clients, such as Bazel, are
setting it in both, newer clients such as Buck2 would only set it in
Action message.

This change replace most of existing command.GetPlatform() calls with
action.GetPlatform() while using the command value as fallback.

This was heavily inspired by https://github.com/buildbuddy-io/buildbuddy/pull/7550, but with a few more changes.

I tried a simpler approach where we overwrite command.platform with action.platform, but because we read the original protos by digest in ExecutionServer.markTaskComplete, this doesn't work.

Closes buildbuddy-io/buildbuddy-internal#3863