Remove os.Stat from disk.EnsureDirectoryExists (#7743)
This stat call isn't necessary, since os.MkDirAll first does a stat by
itself.

This change does affect behavior a bit. Previously, if the path was a
file, the function would return nil. Now it returns an error. As far as
I can tell, this shouldn't matter in practice.

I don't think this will improve performance in dev or prod, since
creating directories seems to be rare. But when running a local server
without any directories, this doubles up the number of stat calls, and
each one occupies a thread, making it more likely that the process dies
because it tries to create more than 10k threads.