[FC] Remove some VBD unmount logging (#7780)
We're suspicious that the failure to unmount VBDs has contributed to
disk buildup on our executors. The logs show that most of the failures
are a result of "context canceled"

By default, we will retry the unmount with an extended context. Remove
logging from the first attempt to unmount, to see if this is actually a
problem in practice