[ci_runner] Support build remotely and running locally via bash command (#7334)
We have a remote bazel feature that builds the target remotely and runs
the executable locally. This happens automatically in the CLI, as
opposed to building the target remotely, and then the user manually
running the executable that was fetched locally (here's the
[PR](https://github.com/buildbuddy-io/buildbuddy/commit/fa9409e55400bd34c19b7e7282d4468999c056fc)
where the behavior was introduced). This PR adds support in the new
Steps syntax - this is the last missing feature before feature parity
with BazelCommands