[Cache Proxy]: attempt to drain the AtimeUpdater queue on server shutdown. (#7829)
This is just a simple first attempt which synchronously round-robins
through groups, sending one atime-updating RPC per group, as long as
there's anything to send. We could do something fancier in the future,
like fire-and-forgetting these RPCs, or adding a cancellation timeout,
but I figured we should just check and see how this does before worrying
about that.