[Cache Proxy] add Cache Proxy Grafana dashboard (#7844)
This PR removes the cache-proxy-specific panels from the BuildBuddy
dashboard and introduces a new Cache Proxy dashboard that's a copy of
the BuildBuddy dashboard with all of the irrelevant stuff removed. We
should switch to some automated method of generating these dashboards
some time, but that's a fair bit more work than this.