Revert prettier upgrade and fix formatting (#7891)
The version upgrade in
https://github.com/buildbuddy-io/buildbuddy/pull/7741 is causing
unexpected prettier diffs whenever we change certain files. This is
because prettier 3.x introduced new formatting rules, but our prettier
script only checks/formats changed files. So, in the upgrade PR,
Checkstyle didn't complain about the formatting now being incorrect
after the upgrade, and the latent formatting errors are being surfaced
in PRs only as the files are edited.

This PR reverts prettier back to 2.1.1 and updates Checkstyle to catch
this error in CI - if the prettier version in yarn.lock changes, then
all files are checked/formatted, not just changed files.