[raft] fix flaky tests (#7895)
1. AddReplica can fail to promote to voter in some rare cases with
request
rejected. When one attempt of `SyncRequestAddReplica` went through in
raft
system, but raft returned a temp "system is busy" error; the next
attempt will
fail with "request rejected". Therefore, when we receive request
rejected error,
instead of returning error directly, we can check the membership to see
if the
replica has been promoted to voter.

2. getMembership in DownReplicate can fail with shard not found between
getting
the store with range lease and getting the membership, the replica was
removed.
In this case, the test should try again to find another store with range
lease
and see the number of replicas on that store.

3. In DownReplicate, sometimes when we check the pebble to see if the
keys are
deleted, the RemoveData hasn't been finished yet. This can happen
because
when we remove replica, we remove it from range descriptor first, then
from
raft; and remove the data in the end. To make the test less flaky, a for
loop
   is added.