[FC] Support invalidating a snapshot by writing a file to the workspace directory (#7958)
Sometimes, firecracker actions are assigned to snapshots that are
corrupted. If the action detects corruption upon startup, it can write
the marker file .BUILDBUDDY_INVALIDATE_RUNNER to the /workspace dir in
the VM. We will invalidate the snapshot, so that on retry, the action
will receive a clean runner. This will also ensure that other actions
aren't assigned to the broken snapshot.

This PR also changes the executor to use the original context when
removing runners, as opposed to the background context. This is needed
because in order to invalidate the snapshot, the group ID must be set on
the context. This should also improve context-based logging during the
remove

https://buildbuddy-corp.slack.com/archives/C0694U6QC30/p1732298274269869