[RB] Fix an incorrect reference to remote branch vs local branch (#8011)
I introduced a bug in this commit:
https://github.com/buildbuddy-io/buildbuddy/pull/7989

If your local branch does not exist remotely, we want to fallback to
using the master branch. In the previous incorrect version, we'd use the
remote master head commit. If this commit didn't exist locally, there'd
be errors when we tried to diff against it to generate the diff
patchsets