Prefer keys with cache write caps (#8020)
In a few different places, we call `GetAPIKeyForInternalUseOnly` to
retrieve an "arbitrary" API key for a group. However, some things break
if the arbitrary key happens to be a readonly key.

This PR makes it so that we retrieve the key with the highest cache
capabilities.

I opted to do ranking instead of filtering, because it seems less risky.
e.g. if an org only has CAS-write or readonly keys, we don't want to
break those orgs.