Write to the action cache in the execution server (#8009)
This is done when the executor sends a completed action to
PublishOperation. The write from the app is added at the same time that
the executor write is removed, so we could have a period of time where
we double write some entries, or don't write some entries, depending on
whether the app or executor rolls out first.