[RB] Fix race condition with fetching completed execution data (#8018)
Fixes
https://buildbuddy-corp.slack.com/archives/C07GMM2VBLY/p1733333761499139?thread_ts=1731949897.755029&cid=C07GMM2VBLY

The CLI sometimes reports the wrong exit code (0, even when the remote
run failed). This is because there can be a slight delay between when
the remote runner execution finishes, and when the execution is
successfully updated in the DB with the final exit code. Make sure to
check that the execution is complete before using the exit code.