Only touch version number when updating bazel module version via sidekick (#8053)
Previously we'd just replace the full module snippet with the example
snippet from the docs.

This could be destructive if either the example snippet includes weird
stuff (like comments), or if usage is non-standard (i.e. has some extra
args).

With this change, we just touch the version argument and bump it to the
latest version when update is clicked.