[raft] RemoveReplica (#8054)
Allow RemoveRequest to take a request without range. If we are removing
zombies,
that means the replica is not in the descriptor, so we don't need the
range
descriptor in that case. But we still want to check that the server is
the lease
holder of that range. 

Known issue: the Cleanup Zombie test can become flaky if we are removing
a leader. I will address it in a follow up PR when I refactor the zombie
code.