Remove workspacefs loopback mount code (#8064)
We aren't actively using this code. The `debugfs` codepath is probably
"fast enough" since we updated it to only copy output paths in
https://github.com/buildbuddy-io/buildbuddy/pull/7959, and also this
code is likely to be supplanted by FUSE anyway. If we do want to try
this code again in the future, we can always bring it back, but should
test thoroughly in dev first because we hit some issues with the
loopback device approach last time.