Don't use COW for workspaces or store them in snapshots (#8068)
It's not necessary to use COW+VBD for the workspace drive or store it as
part of the VM snapshot, because we never wind up reading from the
snapshotted workspace - instead, we wind up just throwing it away after
unpausing, since we create a new workspace ext4 image from the action
workspace on the host for every action.

So, instead of snapshotting the workspace, we can stub out the workspace
drive with an empty file (`empty.bin` relative to the chroot) just
before saving the snapshot, so that the snapshotted state does not
contain any references to the workspace. Firecracker will expect this
file to exist when resuming the snapshot, so we just need to manually
create this file before resuming.

In addition, there is some extra related cleanup in this change: when
initially creating the VM, we currently start out the workspace drive as
an empty ext4 image. The filesystem is empty but the block device
contents contain some bytes for the basic ext4 metadata. The goinit
binary then mounts this filesystem blindly when the VM initially boots.
This was done to ensure that the "setup workspace" routine could always
unmount, swap, and re-mount. However, this is no longer needed. Instead,
we now start out the workspace image as an empty (zero-length) _file_
(not an empty ext4 FS), and don't mount it. This lets us use the same
logic every time for attaching a workspace, regardless of whether we're
creating a new VM, or resuming one from snapshot. In both cases, the VM
starts out with `empty.bin` backing the workspace drive, which is not
mounted. Then, calling `createAndAttachWorkspace` creates an ext4 image
from the host workspace, patches the VM drive contents to reference this
new ext4 file, then mounts the filesystem.