Don't report an AC hit if an output is missing (#8082)
If an output file of an action was missing then we were previously
recording an AC hit even though we'd be returning NotFound to the
client. This PR makes sure we record an AC miss in that case.

Also fixes a bug that when `inline_output_files` is set, we wouldn't
report the correct `ActionResult` response size (we were using
`len(blob)` which doesn't yet incorporate the inlined outputs).