Upgrade `github.com/bazelbuild/bazelisk` to `v1.21.0`
This pulls in https://github.com/bazelbuild/bazelisk/pull/604 to allow
downloading bazel commits on aarch64 linux. The only weird thing about
this is we now provide a repo_mapping to that repo because it started
using the new bzlmod names for rules_go and gazelle. This doesn't
upgrade to the latest version of bazelisk because that would require
other API changes.