fix a bug in tasksize.Requested
I introduced this in https://github.com/buildbuddy-io/buildbuddy/pull/7920 and found it using the new clickhouse fields.

If the user only  sets CPU, it takes no affect. If the user only sets RAM, it sets CPU to 0, which later gets overridden  to 600 from the default.

I fixed this and added tests.