Copy persisted artifacts when replaying an invocation (#8118)
When replaying an invocation, copy the artifacts from blobstore to the
destination cache target. The destination BES handler should then copy
those blobs to its local blobstore, which should allow the artifacts to
be visible in the UI. This is necessary since we recently switched the
/file/download endpoint to avoid reading from cache. I tested this
change before/after and verified it fixes the issue where test logs are
not visible when replaying an invocation.