Add webdriver test for action page (#8117)
The action page breaks often because the `/file/download` logic is
pretty complex as well as the logic in `invocation_action_card.tsx`. So,
it seems worth expanding the existing webdriver test to not just check
the executions listing for the test build, but also click through to the
action page.