Fix page crash error during webdriver tests (#8150)
We are hitting this "session deleted because of page crash" issue:
https://stackoverflow.com/questions/53902507/unknown-error-session-deleted-because-of-page-crash-from-unknown-error-cannot/53970825#53970825

`/dev/shm` inside ociruntime is limited to 64MB (we chose this size to
match podman). Chrome sometimes crashes because this isn't enough
memory.

Passing `--disable-dev-shm-usage` to Chrome makes it stop depending on
`/dev/shm`, which fixes this specific issue.